% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snqr.R
\name{filesToDf}
\alias{filesToDf}
\title{Import data from one or more excel files from SNQ (or elsewhere)}
\usage{
filesToDf(paths, ...)
}
\arguments{
\item{paths}{A character vector of basenames of files}

\item{...}{Arguments forwarded to read_excel}
}
\value{
A data frame
}
\description{
Import data from one or more excel files from SNQ (or elsewhere)
}
