% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snqr.R
\name{findDataPath}
\alias{findDataPath}
\title{Find path where data files are kept}
\usage{
findDataPath()
}
\value{
A character string, either "data/" or "/home/johans/externadata/"
}
\description{
Find path where data files are kept
}
