% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snqr.R
\name{skanesjh}
\alias{skanesjh}
\title{Extract names of townns in Skåne from strings}
\usage{
skanesjh(
  sjh,
  replacements = "Lund|Malmö|Kristianstad|Helsingborg|Halmstad|Växjö|Ystad",
  nas = "övr"
)
}
\arguments{
\item{sjh}{A character vector}

\item{replacements}{A character vector of replacement strings}

\item{nas}{A string to put where none of the replacements fit}
}
\value{
A character vector
}
\description{
Extract names of townns in Skåne from strings
}
