% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snqr.R
\name{uppf_vilkaVar}
\alias{uppf_vilkaVar}
\title{Find and prefix variables from follow-up visits at 2 and 5.5 yrs of age}
\usage{
uppf_vilkaVar(
  uppfdf,
  vilka = c("demo", "2år", "5,5år"),
  brytvar = "Uppföljning skapad av"
)
}
\arguments{
\item{uppfdf}{A data frame of follow-up data from SNQ}

\item{vilka}{A vector of strings specifying one or more of 'demo'graphic variables, '2år' follow up at 2 yrs, '5,5år' follow up at 5.5 yrs}

\item{brytvar}{A string identifying the variable at the start of both blocks of follow-up variables}
}
\value{
A character vector
}
\description{
Find and prefix variables from follow-up visits at 2 and 5.5 yrs of age
}
